Use node-client for API (#137)
Removes a lot of src/messaging and related code as well, perhaps opening up the opportunity for further refactorings in the future. Fix #24. Also updates package-lock, hopefully not breaking anything in the process 😅 

neovim-colorizer is removed as well, since it is no longer in use anywhere (except for `src/components/legacy` and some old tests, which doesn't matter).