no. just. no.
Supposed to be a *working* rewrite of the font-texture-atlas stuff
to use canvaskit so I can make fonts work (have to load from
binary font data gotten from the rust backend, which I can only seem to
do w/canvaskit & not the builtin canvas), but . . . this is just broken on my
1080p monitor (although it looks fine on my MBA's screen), and the code
is just horrifically ugly, messy, convoluted, etc.

The *real* solution here is to just use canvaskit for all (text) rendering
instead of webgl; done right, should be just as fast (canvaskit can use
the GPU), but actually work well for most/all monitors.

In addition, using canvaskit for rendering over the current setup should
allow for adding extra eyecandy/niceties, e.g. Neovide's animated
cursor.

It's probably not gonna be a quick port from webgl -> canvaskit though
(could require a rewrite of both the window-manager etc. code *and*
the rendering code), so I'll just leave it for now until I get the rest
of the UI working as it did pre-port (or about as it did).

So TL;DR plan is leave this broken mess as-is until I get the rest of
the backend ported (w/signature-help etc.), and then refactor/rewrite
the frontend to use canvaskit instead of webgl for rendering, hopefully
cleaning up the terrible mess that currently is src/renderer in the
process.